WIP: UNTESTED: Moved helper scripts into utils directory.
