+++ multiqc/modules/base_module.py
