+++ multiqc/utils/log.py
